// templates/cosmetic/types.ts
export interface ProductColor {
    name: string;
    hex: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  oldPrice?: number;
  imageUrl: string; // Gardé pour rétrocompatibilité (sera l'image principale)
  images: string[]; // Nouveau champ pour la galerie
  discount?: number;
  category: string; // Sous-catégorie
  parentCategory?: string; // Catégorie parente
  promo?: boolean;
  material?: string;
  description?: string;
  quantity: number;
  specifications?: { name: string; value: string; }[];
  colors?: ProductColor[]; // NOUVEAU
  // Nouveau bloc éditorial
  highlights?: {
    title: string;
    imageUrl: string;
    sections: {
      subtitle: string;
      features: {
        title: string;
        description: string;
      }[];
    }[];
  };
}